repack_kernel_boot_images: Add a script to repack local kernel build boot images
This script is intended to avoid the dependency issue between device
base-rom and kernel prebuilt ramdisks under
prebuilt/boot-artifacts/ramdisk/. To achieve this, this script will
dump/extract ramdisks from device via ADB and recreate boot.img and
vendor_boot.img based on it.

This script does not re-create any of the vbmeta partition images which
dm-verity has to be disabled prior to flashing these images.

Bug: 203499549
Signed-off-by: Robin Peng <robinpeng@google.com>
Change-Id: Ib911609bb090ae6c0800b4667211c731055ff653