usb: host: unregister the usb notifier if probe() fail
When xhci-exynos probe failed, the double register issue
will happen for the second time probe of xhci-exynos driver.
Currently the unregister is only called when xhci-exynos is removed,
we should add the unregister when probe fails.

The patch will add error handling to unregister the notifier
during probe to fix this issue.

Bug: 208386770
Signed-off-by: Ray Chi <raychi@google.com>
Change-Id: I568d4613a47e19da96199eb7b2da802811092a74