repack_kernel_boot_images: Add prompt indicate user to redownload or reuse ramdisks
Once the script dump ramdisk stuff from device and generated images, the
repacked_dir will have the synced ramdisk from the device. Suppose we
can reuse it if device remain with the same baseo-rom. Add a prompt let
user to decide if they like to redownload partition or reuse the
previous dumped ramdisks.

Bug: 203499549
Signed-off-by: Robin Peng <robinpeng@google.com>
Change-Id: Idd272598cf1ca0b1be496c0f23d77f0eb96879fe