update_symbol_list.sh: fix abi xml generation
To determine if we need to generate an ABI xml update, the script checks
to see if the new symbols added are in any of the other symbol lists.
In order to check the symbol differences with `git diff ACK..HEAD`, we
need to first commit the changes so that HEAD reflects the symbol diff.
This patch refacts the commit generation logic to commit the symbol list
changes first before checking if an ABI xml update is needed

Bug: 210910786
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ib48c96fdb3216077d29d8f68831a962e107884a8