vendor_hook: sched: Implement vendor group util
Cpu cgropus provide per-group util, however, it will combine with cpu.shares.
If cpu.shares is not enabled, then we could not use per-group util. Besides,
it does not provide the estimated util per group. This patch implements
the vendor per-group util as well as per-group estimated util. With
this, we could control the max util per group for power consideration.

Bug: 201260585
Signed-off-by: Rick Yiu <rickyiu@google.com>
Change-Id: I0eeee1a4960bfbb654b2d79549dea25fceae2c75