pogo_transport: Allow enabling data over USB-C while being docked
This change introduces a device tree node "equal-priority" which
when set treats both data transfer over USB-C and POGO to be of
equal priority. This implies that when usb data is already bounded
to either USB-C or POGO, it would not be rebounded to the one arriving
later till the current bound client has released the usb data path.
So, "equal-priority" allows USB-C or POGO to grab usb data on a
first come first serve basis.

The change creates the below device attributes:
1. pogo_docked: Reads 1 when docked
2. pogo_usb_capable: Reads 1 when pogo link is capable of usb data
3. pogo_usb_active: Reads 1 when usb data is enabled for pogo path
4. equal_priority: Reads 1 when both USB-C and pogo are treated
                   with equal priority
5. move_data_to_usb: Write 1 to enable USB data over USB path.
                     This would unbind USB Data from the pogo path
                     during docked state.

Bug: 211677613
Bug: 213312081
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
Change-Id: I58be1e1f34c4fb2bee2660fa32b1481893d421f9