arm64/dts: gs201: Disable the AoC & USB PPMU's by default
When the AoC power domain is down, accessing the AoC & USB PPMU's
causes an ITMON crash. Currently Debug Core is not being updated about
the state of this power domain, but there are global BCM commands that
access all PPMU's at once. These commands rely on the power domain
information to avoid PPMU's that are not powered, but this fails for
AoC & USB - therefore, these PPMU's are disabled by default.

Bug: 215569471
Signed-off-by: Peter Csaszar <pcsaszar@google.com>
Change-Id: Ibc620beb89b3d0db91632e8124deaa6ba038d8a3