Refine the check of pmu limit condition
The cpu frequency is dominated by the cpu with highest util of the
cluster. We don't need every cpu in the policy to pass the check of
pmu limit condition to make the freq limit effective. We just need to
make sure if a cpu failed the check, its expected freq does not exceed
the limit freq.

Bug: 170511572
Signed-off-by: Rick Yiu <rickyiu@google.com>
Change-Id: I786741c10bf4c0e815bdcdd29b7c4c8cd6966648