iommu/samsung: Fix leak of lv2 page tables in .unmap_pages
The .unmap_pages callback mistakenly assumed that the page size passed
in is always identical with the one that has been passed into the map
function. However, the formula to calculate the optimal page size in the
unmap code path is different from the map code path. We encountered
situations where the .unmap_pages callback was called with a page size
of 1MB when the actual page size used in the mapping path was smaller
than this. The code then erroneously assumed that the first level page
table entry is a section mapping when it in fact was a pointer to a
second level page table. It then just set the first level page table
entry to 0 effectively erasing the pointer to the second level page
table and therefore leaking the memory allocation that had been reserved
for the second level page table.

Bug: 219580120
Fixes: 847978d78b3a ("iommu/samsung: optimize PTE flush")
Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: I262453b38f18bba280f9867dcf0a30387e33e706