pkvm-s2mpu: Print FMPT/SMPT registers in IRQ handler
When an S2MPU triggers an IRQ, we print information about the fault. It
can be useful to know more about the state of the S2MPU at that point,
so print the values of L1ENTRY_ATTR and L1ENTRY_L2TABLE registers for
the VID of the fault.

Bug: 229793579
Signed-off-by: David Brazdil <dbrazdil@google.com>
Change-Id: Iec758355b7611943f49c4f45384b2bba9f7443fc