vendor_hook: sched: Change the uclamp max of low prio tasks
The cpu importance value is the sum of its uclamp max and min, and
due to bucket design, if there is long running low prio task, then
the cpu uclamp max and min will not change. So, lower the uclamp
max of low prio task to one bucket in order to make cpu importance
work in this condition.

Bug: 228902039
Signed-off-by: Rick Yiu <rickyiu@google.com>
Change-Id: Iaefa47bc96305d02cee3870c9bf060c7730fe0a7