google/debug: ehld: handle hrtimers correctly in CPUHP and CPUPM
Fix hrtimer handling in CPUHP and CPUPM routines.

CPUHP cpu start is where the EHLD hrtimer is first started.
CPUHP cpu stop is where the EHLD hrtimer finally stops.

While the core is online:
Stop the EHLD hrtimer in CPUPM enter, so that the core doesn't have to
unnecessarily wake up just because of a running EHLD hrtimer. Let the
core sit in its PM peace. Debug Core is aware of that too.

Restart the hrtimer in CPUPM exit, when the core gets ready to run code.

There are many unexpected call sequences between CPUHP and CPUPM.
Manual testing shows that the following are possible:

1) CPUHP cpu start
   CPUPM exit
   CPUPM enter
   --
   CPUHP cpu start starts the hrtimer for that core.
   CPUPM exit should not try to restart the same hrtimer.

2) CPUHP cpu stop
   (CPU comes back online)
   CPUPM enter
   CPUPM exit
   CPUHP cpu start
   --
   CPUPM calls before CPUHP cpu start need to be ignored.

3) CPUHP cpu stop
   CPUPM enter
   CPMPM exit
   --
   CPUPM calls after CPUHP cpu stop need to be ignored.

ctrl->ehld_running is used to track CPUHP state.
ctrl->ehld_cpupm is used to track CPUPM state.

Bug: 228981355
Bug: 231535301
Signed-off-by: Petri Gynther <pgynther@google.com>
Change-Id: I4681c4df6a9cabb5e9d74137b488d6b9ce908e59