Revert^4: s2mpu: Parse and register SYSMMU_SYNC devices
PRO B0 chip adds new SYSMMU_SYNC devices that TZ/EL1/EL2 can use to
wait until all in-flight DMA transactions of a given master have
drained. This is used to confirm that page-table invalidation has
completed. pKVM exports pkvm_iommu_sysmmu_sync_register() to register
a SYSMMU_SYNC device with the hypervisor.

Iterate over a list of phandles in the 'sysmmu_syncs' DT property of
each S2MPU. Find the corresponding device and register it with pKVM.
This is implemented inside the pkvm-s2mpu.ko, without creating a
separate driver for SYNCs, in order to save runtime memory.

It is not necessary to invoke pkvm_iommu_suspend/resume on SYNC devices.
The hypervisor assumes they follow the power transitions of the parent
S2MPU. This is the case for all SYNCs in PRO.

This reverts commit 8f60204ba2c7fcd51f39c073da6b080e62a6bdf7.

Bug: 233147450
Test: AndSys SuspendTest
Signed-off-by: David Brazdil <dbrazdil@google.com>
Change-Id: I859e6895e40b4cd71a4ea78cba74f4811e0bce3e