bcl: Ensure correct read of temperature reading
Whenever IRQ is triggered, temperature reading is updated.  However, due
to transient nature of IRQ, temperature reading often gets reset before
the triggered value can be read and processed by upstream consumer.  To
mitigate this, we prolong the triggered value by adding a timer to
prevent reset of the value.  It is noted that it takes approximately
50ms for upstream consumer to consume the data.  Timer is set to 50ms.

Bug: 238555356
Test: Ensuring proper temperature value is read from userspace during
triggering.
Signed-off-by: George Lee <geolee@google.com>
Change-Id: Id5a44fb213399ab6aa04c07c0e7312c619e1108d