pcie: exynos: Get ch_num based on vendor id
The previous code made an implied assumption that the modem pcie
interface will precede the wifi pcie interface. However if the probing
order gets switched then this causes an issue with how the data
structures for the pcie interfaces are retrieved. This code change
checks the endpoint vendorid to select the appropriate pcie interface
index.

Bug: 246251383
Change-Id: I33251ff1fdac449b69801a46f801aaa1b3fbf558
Signed-off-by: Sajid Dalvi <sdalvi@google.com>