media: mfc: support the BT.601_525 and BT.2020(8bit)
This patch supports the option to use the coefficients of
the CSC formula determined by firmware on COLOR_SPACE
and COLOUR_PRIMARIES of E_VIDEO_SIGNAL_TYPE.
With this option, BT.601-525 and BT.2020(8bit) are available.
And VUI is always set regardless of standard to use
new option "enc_rgb_csc_by_fw".

Bug: 239912381
Signed-off-by: AyoungSim <a.sim@samsung.com>
Signed-off-by: wenchangliu <wenchangliu@google.com>
Change-Id: I768a72ed9ae89d93f8a729854cfcf3ad78766b85