vh: cgroup: Bypass rebuild_root_domains when system resume
In rebuild_root_domains, it loops every task in the system to update
dl bandwidth if there is dl task, which is time consuming, even there
is no dl task in the system. It affects system resume time because
it is in the critical path, so skip it since there is no dl task in the
system for now.

Bug: 238390134
Change-Id: I49aca77c61bba30c74ce5c12651fcc4b89ab5ad5
Signed-off-by: Rick Yiu <rickyiu@google.com>