Revert "Change device tree for P22 to insulate UWB from AoC crashes"
Revert submission 2311369-b/240205717

Reason for revert: b/259450077 - This change is causing CCC not to work but the change should be unrelated.  The thought is that this change is changing timing and the qorvo hal isn't handling this correctly.  Plan is to back out this change until we root cause the issue.

Reverted Changes:
I723e8af9a:Modify dw3000 driver to insulate UWB from Aoc
I10ed786d3:Change device tree for P22 to insulate UWB from Ao...

Signed-off-by: Victor Liu <victorliu@google.com>
Change-Id: I596afb0b1f6a8418e11c28c35d44c5290291f472