usb: dwc3: delete the timer after configured state
Basically, the recovery method is triggered by a timer, and it
will exit directly if the gadget state is configured. However,
when the method is checking the gadget state and the state is
changed from configured to disconnected, the recovery will be
running. It will cause some ITMON issues if recovery is running
in disconnected scenario.

This patch will delete the timer if the gadget state is configured
so that the recovery wouldn't be triggered.

Bug: 263435622
Change-Id: I816163e4148a8c569902be7cc6e36ec386c4faf4
Signed-off-by: Ray Chi <raychi@google.com>