PIXEL: ufs: fips: move access to encryption functionality to fips module
FIPS140-3 requires the CMVP module to control access to encryption
services. Encryption services are controlled via the Crypto Enable
flag in the PRDT. This CL moves the functionality to construct the
PRDT away from the core driver and into the CMVP module when the
CMVP module is enabled in the Kernel.

Bug: 263177403
Change-Id: I8edcd9bace0be572b3f4dae3d3d05d196b9cc7a4
Signed-off-by: Konstantin Vyshetsky <vkon@google.com>