google/debug: etm2dram: Introduce driver to control etm2dram
Exposes interfaces to control debugcore plugin 'etm2dram' - the plugin
maintains an internal timer that can be configured to capture ETM to
DRAM when times out. So the interface's user is supposed to set the
timeout repeatedly, and the capturing will be started after that user
stops petting the timer, which is usually when the system is out of
response. Once the system is recovered by a system fault handler that
able to capture memory dump, the data captured by ETM will be available
in the dump for the further debugging.

Bug: 265382296
Change-Id: If6435ffadef846a817d526cbf3f123b1b7b4249c
Signed-off-by: Woody Lin <woodylin@google.com>