pcie: exynos: Add check to ensure pcie phy regs are reset
This is a temporary change to determine if the APC watchdogs seen on P22
are related to the PCIe PHY not being reset properly. This code checks a
few registers to see if they match up to their reset values. In the APC
watchdog the code seems to lock up while trying to write to PCIe PHY
address 0xD8. The reads added here can also help determine if the issue
is only seen on a write or also on read.
Once we get some failures with this change we can revert it.

Bug: 266855627
Change-Id: I69b29c7ffb87027158041f3d81e2f9d085b8fd4a
Signed-off-by: Sajid Dalvi <sdalvi@google.com>