arm64/s2mpu: Add new DT property dma-cons.
dma-cons is used for devices that are actively using DMA while
device is booting, and if the S2MPU is probed during the device it
can enable runtime PM for it and as the device has not reference, it
will be suspend causing S2MPU faults.

This patch makes the probe function get a reference for devices with
dma-cons property if they have a power domain. It will also link the
s2mpu(suppiler) to the DMA master referenced in dma-cons(consumer).

When other device probe is done sync_state() would be called which
would put the reference for the S2MPU if it was get at probe.

Bug: 262187119
Change-Id: I7f2c1747a176536352a7db5d23bb748f1d9003c4
Signed-off-by: Mostafa Saleh <smostafa@google.com>