gki: Select minimal set of GKI hack configs for cloudripper
There are only a few configs which are dependencies of cloudripper modules
but need to be built-in rather than built as modules. Select the minimal
set of these configs that are actually required for cloudripper in order to
cut out a lot of GKI cruft.

This also removes Kconfig.gki from the Kconfig system to make menuconfig
look tidy again.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>