media: mfc: support DRC cancel
If the streamoff() calls during DRC processing,
driver has been waiting for DRC processing to be completed.
However, since DRC processing may not be completed for
various reasons(there is no dst buffer or no interrupt...),
driver cancel the DRC when there is a streamoff() during
DRC processing.
Driver stops the LAST_FRAME command with the F/W, processes
streamoff() and cleanup all buffers.
In order to proceed from SEQ_START with the new input buffer
that came in later, DRC flag is delivered to the user and
a new sequence is started.

Bug: 269568991
Change-Id: I8cb13d7eba68b92a6184eed6ce0a1d2854544108
Signed-off-by: AyoungSim <a.sim@samsung.com>