vh: sched: Fix iowait boost updates
Commit 44070558c0071 ("vh: sched: Block cpufreq updates until uclamp
filters are applied") accidently broke iowait boost as it forgot to
pass on the flag when we temporarily block cpufreq updates. Which meant
when we go back and resume, the iowait update isn't applied causing some
regression on workloads that are sensitive to this parameter.

Rectify it by passing the iowait boost flag so that we can remember to
apply the boost when we resume the cpufreq updates.

Bug: 279805535
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Id1949cb88e636cf2e916fe0be5866c99c84dfead