vh: sched: Don't select the previous CPU if it isn't allowed
The previous CPU isn't always allowed anymore, so waking a task on it when
it's not allowed is a major bug. The scheduler doesn't check if the CPU
returned by select_task_rq() is allowed, so it honors invalid CPUs chosen
by select_task_rq().

Fix the illegal task migrations by checking if the waking task is actually
permitted to use its previous CPU, and using a different allowed CPU if so.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>